split out the analysis objects
