+++ server/opendp_apps/analyses/__init__.py
@@ -1,0 +1,2 @@
+
+#default_app_config = 'analyses.apps.AnalysesConfig'
