+++ server/opendp_apps/analyses/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class AnalysesConfig(AppConfig):
+    name = 'analyses'
+    #verbose_name = 'DP Analysis'
+    #verbose_name_plural = 'DP Analyses'
