+++ server/opendp_apps/analyses/models.py
@@ -1,93 +1,9 @@
-from django.core.files.storage import FileSystemStorage
 from django.db import models
-from django.db.models import CASCADE
 from django.conf import settings
-from polymorphic.models import PolymorphicModel
+from opendp_apps.datasets.models import DataSetInfo
 from opendp_apps.model_helpers.models import \
-    (TimestampedModel, TimestampedModelWithUUID)
+    (TimestampedModelWithUUID,)
 
-UPLOADED_FILE_STORAGE = FileSystemStorage(location=settings.UPLOADED_FILE_STORAGE_ROOT)
-
-
-class DataSetInfo(TimestampedModelWithUUID, PolymorphicModel):
-    """
-    Base type for table that either holds DV data
-    or a file upload
-    """
-    name = models.CharField(max_length=128)
-
-    # user who initially added/uploaded data
-    creator = models.ForeignKey(settings.AUTH_USER_MODEL,
-                                on_delete=models.PROTECT)
-
-    class SourceChoices(models.TextChoices):
-        UserUpload = 'upload', 'Upload'
-        Dataverse = 'dataverse', 'Dataverse'
-
-    source = models.CharField(max_length=128,
-                              choices=SourceChoices.choices)
-
-    # Redis key to store potentially sensitive information
-    # during analysis setup
-    data_profile_key = models.CharField(max_length=128, blank=True)
-
-    class Meta:
-        verbose_name = 'Dataset Information'
-        verbose_name_plural = 'Dataset Information'
-        ordering = ('name', '-created')
-
-    def __str__(self):
-        return self.name
-
-
-class DataverseFileInfo(DataSetInfo):
-    """
-    Refers to a DV file from within a DV dataset
-    """
-    # TODO: This should have all fields from DV API response
-    dataverse_file_id = models.CharField(max_length=128)
-    doi = models.CharField(max_length=128)
-    installation_name = models.CharField(max_length=128)
-
-    class Meta:
-        verbose_name = 'Dataverse File Information'
-        verbose_name_plural = 'Dataverse File Information'
-        ordering = ('name', '-created')
-
-    def __str__(self):
-        return f'{self.name} ({self.installation_name})'
-
-    def save(self, *args, **kwargs):
-        # Future: is_complete can be auto-filled based on either field values or the STEP
-        #   Note: it's possible for either variable_ranges or variable_categories to be empty, e.g.
-        #       depending on the data
-        #
-        self.source = DataSetInfo.SourceChoices.Dataverse
-        super(DataverseFileInfo, self).save(*args, **kwargs)
-
-
-
-class UploadFileInfo(DataSetInfo):
-    """
-    Refers to a file uploaded independently of DV
-    """
-
-    # user uploaded files, keep them off of the web path, etc.
-    #
-    data_file = models.FileField('User uploaded files',
-                    storage=UPLOADED_FILE_STORAGE,
-                    upload_to='user-files/%Y/%m/%d/')
-
-    def save(self, *args, **kwargs):
-        # Future: is_complete can be auto-filled based on either field values or the STEP
-        #   Note: it's possible for either variable_ranges or variable_categories to be empty, e.g.
-        #       depending on the data
-        #
-        self.source = DataSetInfo.SourceChoices.UserUpload
-        super(UploadFileInfo, self).save(*args, **kwargs)
-
-    def __str__(self):
-        return f'{self.name} ({self.source})'
 
 
 class DepositorSetupInfo(TimestampedModelWithUUID):
@@ -119,7 +35,7 @@ class DepositorSetupInfo(TimestampedModelWithUUID):
     def __str__(self):
         return f'{self.dataset} - {self.user_step}'
 
-    def save(self):
+    def save(self, *args, **kwargs):
         # Future: is_complete can be auto-filled based on either field values or the STEP
         #   Note: it's possible for either variable_ranges or variable_categories to be empty, e.g.
         #       depending on the data
@@ -138,6 +54,9 @@ class AnalysisPlan(TimestampedModelWithUUID):
         STEP_020_CUSTOM_VARIABLES = 'step_020', 'Step 20: Custom Variables'
         STEP_030_VARIABLE_TYPES = 'step_030', 'Step 30: Confirm Variable Types'
         STEP_100_ANALYSIS_READY = 'step_100', 'Step 100: Analysis ready!'
+
+    analyst = models.ForeignKey(settings.AUTH_USER_MODEL,
+                                on_delete=models.PROTECT)
 
     name = models.CharField(max_length=255)
 
@@ -170,7 +89,7 @@ class ReleaseInfo(TimestampedModelWithUUID):
     dataset = models.ForeignKey(DataSetInfo,
                                 on_delete=models.PROTECT)
 
-    # redundant
+    # also gives analyst
     analysis_plan = models.ForeignKey(AnalysisPlan,
                                       on_delete=models.PROTECT)
