--- server/opendp_apps/analysis/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class AnalysisConfig(AppConfig):
-    name = 'analysis'
