--- server/opendp_apps/dataset/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class DataverseConfig(AppConfig):
-    name = 'dataverse'
