+++ server/opendp_apps/datasets/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class DatasetsConfig(AppConfig):
+    name = 'datasets'
